+++ app/bridge3/app/domain/types/wallet.ts
@@ -1,9 +1,7 @@
-import React from 'react'
-
 interface Wallet {
   address: string | `0x${string}`
   providerName: string
-  icon: (props: any) => React.ReactNode
+  icon: (props: any) => JSX.Element
   connector?: string
   metadata?: any
   chainId?: string | number
