--- app/bridge3/app/domain/wallet/providers/index.ts
@@ -1,1 +1,0 @@
-export { default as useEvmProvider } from './use-evm-provider'
