+++ app/bridge3/app/luxkit/evm/resolve-evm-icon.tsx
@@ -1,9 +1,9 @@
-import MetaMaskIcon from '@/components/icons/wallets/metamask'
-import WalletConnectIcon from '@/components/icons/wallets/walletconnect'
-import BitKeep from '@/components/icons/wallets/bitkeep'
-import RainbowIcon from '@/components/icons/wallets/rainbow'
-import CoinbaseIcon from '@/components/icons/wallets/coinbase'
-import Phantom from '@/components/icons/wallets/phantom'
+import MetaMaskIcon from '../icons/wallets/metamask'
+import WalletConnectIcon from '../icons/wallets/walletconnect'
+import BitKeep from '../icons/wallets/bitkeep'
+import RainbowIcon from '../icons/wallets/rainbow'
+import CoinbaseIcon from '../icons/wallets/coinbase'
+import Phantom from '../icons/wallets/phantom'
 
 import { Coins, type LucideProps } from 'lucide-react'
 import { type Connector } from 'wagmi'
