+++ app/bridge3/app/luxkit/index.ts
@@ -3,3 +3,6 @@ export {
   LuxEvmContext, 
   default as LuxEvmProvider
 } from './evm'
+export {
+  default as useWallets
+} from './use-wallet'
