+++ app/bridge3/app/luxkit/use-wallet.ts
@@ -2,8 +2,7 @@ import { toast } from '@hanzo/ui/primitives-common'
 
 import type { Network } from '@luxfi/core'
 import type { SwapItem, Wallet, WalletProvider } from '@/domain/types'
-
-import { useEvmProvider } from './providers'
+import { useLuxEvm } from '.'
 
 interface useWalletsReturn {
   wallets: Wallet[]
@@ -11,13 +10,12 @@ interface useWalletsReturn {
   disconnectWallet: (providerName: string, swap?: SwapItem) => Promise<void>
   getWithdrawalProvider: (network: Network) => WalletProvider | undefined
   getAutofillProvider: (network: Network) => WalletProvider | undefined
-  getAutofillProviderUsingName: (network: string) => WalletProvider | undefined
 }
 
 const useWallets = (): useWalletsReturn => {
 
   const WalletProviders: WalletProvider[] = [
-    useEvmProvider(),
+    useLuxEvm(),
     // useTON(),
     // useStarknet(),
     // useImmutableX(),
@@ -36,12 +34,6 @@ const useWallets = (): useWalletsReturn => {
   const handleDisconnect = async (providerName: string, swap?: SwapItem): Promise<void> => {
     const provider = WalletProviders.find((provider) => provider.name === providerName)
     try {
-      // if (swap?.source_exchange) {
-      //   const apiClient = new BridgeApiClient()
-      //   await apiClient.DisconnectExchangeAsync(swap.id, 'coinbase')
-      // } else {
-      //   await provider?.disconnectWallet()
-      // }
       await provider?.disconnectWallet()
     } 
     catch {
@@ -61,7 +53,7 @@ const useWallets = (): useWalletsReturn => {
   }
 
   const getWithdrawalProvider = (network: Network) => {
-    const provider = WalletProviders.find((provider) => provider.withdrawalSupportedNetworks.includes(network.internal_name))
+    const provider = WalletProviders.find((provider) => provider?.withdrawalSupportedNetworks?.includes(network.internal_name))
     return provider
   }
 
@@ -71,19 +63,12 @@ const useWallets = (): useWalletsReturn => {
     return provider
   }
 
-  const getAutofillProviderUsingName = (network: string) => {
-    console.log(WalletProviders)
-    const provider = WalletProviders.find((provider) => provider?.autofillSupportedNetworks?.includes(network))
-    return provider
-  }
-
   return {
     wallets: getConnectedWallets(),
     connectWallet: handleConnect,
     disconnectWallet: handleDisconnect,
     getWithdrawalProvider,
     getAutofillProvider,
-    getAutofillProviderUsingName,
   }
 }
